Rename and reorganise into package (#2)
* Switch configurable RPi GPIO when authenticated

If a valid NFC tag is detected, the configured GPIO is toggled.
If the GPIO is toggled from low to high, a scheduler is started to set the pin low after 2 hours.
If the GPIO is toggled from high to low, the scheduler is stopped.

* Add requirements.txt

* Add IDE configuration

* Create package nfclock

* Add tag release callback

If the NFC tag is removed from the reader, wait 10 seconds
before switching off the configured GPIO

* Reduce IDE warnings

* Make sqlite path configurable

* Add setup.py

* Add gitignore

* Update deployment settings

* Add scheduler thread

* Add reader thread

* Integrate scheduler and reader threads into main module

This commit also changes nfclocks behaviour. The configured pin is kept high as
long as the NFC tag remains in communications range of the reader. After the tag
 is removed, nfclock waits an additional 10 seconds before it pulls the
 configured pin low.

* Update README.md

Co-authored-by: Steffen Arntz <winrootkit@gmail.com>